requirements.txt
loguru==0.7.2
numpy==2.1.0
pandas==2.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
